move format to formatter.ts
