+++ src/renderers/index.ts
@@ -1,6 +1,5 @@
 import html from './html';
 import react from './react/react';
 import reactStatic from './react/static';
-import UNSTABLE_DO_NOT_USE_pretty_render from './unstable_pretty';
 
-export default { html, react, reactStatic, UNSTABLE_DO_NOT_USE_pretty_render };
+export default { html, react, reactStatic };
